fix: Fix 9 server-side game engine bugs
Plan: .claude/plans/snug-enchanting-reddy.md

Fix 1 (Critical): ServerHand RAISE/BET set currentBet to addend, not
running total, breaking isPotGood() and causing infinite re-prompts.
Fix 2 (High): initPlayerIndex() had unreachable break guard, causing
infinite loop when all players are all-in.
Fix 3 (Medium): GameInstanceManager.createGame() TOCTOU race fixed
with synchronized.
Fix 4 (Medium): Sitting-out players now auto-fold via processPlayerAction
instead of no-op setSittingOut(true).
Fix 5 (Medium): ServerGameTable.isCurrent() now returns true.
Fix 6 (Medium): Addon offer guarded against addonLevel = -1 (disabled).
Fix 7 (Minor): Silent long→int profileId casts replaced with
Math.toIntExact() in GameInstance and InboundMessageRouter.
Fix 8 (Minor): GameEventBroadcaster PlayerActed case now bounds-checks
tableId before lookup.
Fix 9 (Minor): PlayerActed and ShowdownPlayerData now populate player
name from seat lookup instead of empty string.

+++ code/pokergameserver/src/main/java/com/donohoedigital/games/poker/gameserver/websocket/InboundMessageRouter.java
@@ -290,7 +290,7 @@ public class InboundMessageRouter {
             sendError(connection, "INVALID_DATA", "Invalid rebuy decision data");
             return;
         }
-        game.submitRebuyDecision((int) connection.getProfileId(), data.accept());
+        game.submitRebuyDecision(Math.toIntExact(connection.getProfileId()), data.accept());
     }
 
     private void handleAddonDecision(PlayerConnection connection, GameInstance game, JsonNode dataNode) {
@@ -301,7 +301,7 @@ public class InboundMessageRouter {
             sendError(connection, "INVALID_DATA", "Invalid addon decision data");
             return;
         }
-        game.submitAddonDecision((int) connection.getProfileId(), data.accept());
+        game.submitAddonDecision(Math.toIntExact(connection.getProfileId()), data.accept());
     }
 
     private PlayerAction parseAction(String actionString, int amount) {
